+++ Gymgenius/Gymgenius/BLL/ExerciseToProgramManagment.cs
@@ -10,25 +10,50 @@ namespace GymGenius.BLL
     public class ExerciseToProgramManagment
     {
         private readonly IExerciseToProgramRepository _exerciseToProgram;
+        private readonly IExerciseRepository _exercise;
+        private readonly ITrainingProgramRepository _training;
 
-        public ExerciseToProgramManagment(IExerciseToProgramRepository exerciseToProgram)
+        public ExerciseToProgramManagment(IExerciseToProgramRepository exerciseToProgramRepository, IExerciseRepository exerciseRepository, ITrainingProgramRepository trainingProgramRepository)
         {
-            _exerciseToProgram = exerciseToProgram;
+            _exerciseToProgram = exerciseToProgramRepository;
+            _exercise = exerciseRepository;
+            _training = trainingProgramRepository;
         }
 
-        public void AddExerciseToProgram(Exercise exercise, TrainingProgram program)
+        public void AddExerciseToProgram(string exerciseName, int programId)
         {
+            Exercise exercise = _exercise.GetExerciseByName(exerciseName);
+            TrainingProgram program = _training.GetTrainingProgramById(programId);
+
+            if (_exerciseToProgram.IsExerciseExistsInProgram(exercise, program))
+            {
+                throw new Exception("Exercise already in program.");
+            }
+
             _exerciseToProgram.AddExerciseToProgram(exercise, program);
         }
 
-        public void DeleteExerciseFromProgram(Exercise exercise, TrainingProgram program)
+        public void DeleteExerciseFromProgram(string exerciseName, int programId)
         {
+            Exercise exercise = _exercise.GetExerciseByName(exerciseName);
+            TrainingProgram program = _training.GetTrainingProgramById(programId);
+
+            if (!_exerciseToProgram.IsExerciseExistsInProgram(exercise, program))
+            {
+                throw new Exception("Exercise not in program.");
+            }
+
             _exerciseToProgram.DeleteExerciseFromProgram(exercise, program);
         }
 
-        public List<Exercise> GetAllExerciseOfProgram(TrainingProgram program)
+        public List<Exercise> GetAllExerciseOfProgram(int programId)
         {
-            return _exerciseToProgram.GetAllExerciseOfProgram(program);
+            if (!_training.IsTrainingProgramExists(programId))
+            {
+                throw new Exception("No program found.");
+            }
+
+            return _exerciseToProgram.GetAllExerciseOfProgram(_training.GetTrainingProgramById(programId));
         }
 
         public bool IsExerciseExistsInProgram(Exercise exercise, TrainingProgram program)
