+++ Gymgenius/Gymgenius/BLL/UserToProgramManagment.cs
@@ -1,4 +1,5 @@
 ﻿using Gymgenius.bo;
+using Gymgenius.dal;
 using GymGenius.BO;
 using GymGenius.DAL;
 
@@ -7,25 +8,40 @@ namespace GymGenius.BLL
     public class UserToProgramManagment
     {
         private readonly IUserToProgramRepository _userToProgramRepository;
-
-        public UserToProgramManagment(IUserToProgramRepository userToProgramRepository)
+        private readonly IUserRepository _userRepository;
+        private readonly ITrainingProgramRepository _trainingProgramRepository;
+        public UserToProgramManagment(IUserToProgramRepository userToProgramRepository, IUserRepository userRepository, ITrainingProgramRepository rainingProgramRepository)
         {
             _userToProgramRepository = userToProgramRepository;
+            _userRepository = userRepository;
+            _trainingProgramRepository = rainingProgramRepository;
         }
 
-        public void AddProgramToUser(User user, TrainingProgram program)
+        public void AddProgramToUser(int userId, int programId)
         {
+            User user = _userRepository.GetUserById(userId);
+            TrainingProgram program = _trainingProgramRepository.GetTrainingProgramById(programId);
+            if (_userToProgramRepository.IsUserHaveProgram(user))
+            {
+                throw new Exception("User already have program.");
+            }
+
             _userToProgramRepository.AddProgramToUser(user, program);
         }
 
-        public TrainingProgram GetUserProgram(User user)
+        public TrainingProgram GetUserProgram(int userId)
         {
-
-            return _userToProgramRepository.GetUserProgram(user);
+            return _userToProgramRepository.GetUserProgram(_userRepository.GetUserById(userId));
         }
 
-        public void RemoveProgramFromUser(User user)
+        public void RemoveProgramFromUser(int userId)
         {
+            User user = _userRepository.GetUserById(userId);
+            if (!_userToProgramRepository.IsUserHaveProgram(user))
+            {
+                throw new Exception("No program found for user.");
+            }
+
             _userToProgramRepository.RemoveProgramFromUser(user);
         }
